refactor(pty): extract OscResponder from TerminalProcess
- Move OSC 10/11 color-query regex constants and response logic into
  electron/services/pty/OscResponder.ts as a stateless exported function
  matching the headlessResponder collaborator pattern.
- Replace the 25-line inline block in TerminalProcess.setupPtyHandlers.onData
  with a single call; the shouldHandleOscColorQueries gate and the
  data.includes("\x1b]1") fast-path stay at the call site.
- Write errors are swallowed silently inside the helper, mirroring
  headlessResponder; the strip-on-success invariant preserves the fail-open
  behaviour the existing tests pin (15 OSC tests still green).

Closes #5925
